requirements.txt
asgiref @ file:///croot/asgiref_1724072470684/work
Django @ file:///croot/django_1731615954020/work
django-jalali==7.4.0
django-jalali-date==2.0.0
djangorestframework==3.16.0
jalali_core==1.0.0
jdatetime==5.2.0
pillow==11.1.0
sqlparse @ file:///croot/sqlparse_1734704446487/work



# asgiref==3.8.1
# Django==5.1.3
# pillow==11.1.0
# pip==25.0
# setuptools==75.8.0
# sqlparse==0.5.2
# wheel==0.45.1